refactor: abstracted validation to separate class
